+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 Improvements:
 * [ODRC-24](https://openlmis.atlassian.net/browse/ODRC-24) Global header and translations implemented for reports
 New functionality added in a backwards-compatible manner:
+* [OPSD-45](https://openlmis.atlassian.net/browse/OPSD-45): Allow cancelling an order that cannot be fulfilled instead of confirming a blank shipment.
 * Added extension Flyway migration support - extensions can now ship their own SQL migrations in `db/extension/`, tracked independently in a separate `extension_schema_version` table
 * [ODRC-66](https://openlmis.atlassian.net/browse/ODRC-66) Update VVM status validator